@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class AddProfileFieldsToUsers < ActiveRecord::Migration
+class AddProfileFieldsToUsers < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :username, :string
     add_index :users, :username, unique: true
