@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreateArticles < ActiveRecord::Migration
+class CreateArticles < ActiveRecord::Migration[4.2]
   def change
     create_table :articles do |t|
       t.string :title
