@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreateFavorites < ActiveRecord::Migration
+class CreateFavorites < ActiveRecord::Migration[4.2]
   def change
     create_table :favorites do |t|
       t.references :user, index: true, foreign_key: true
